+++ Runtime/Scripts/Export/HighDefinitionMaterialExport.cs
@@ -17,6 +17,7 @@
 
 using System;
 using GLTFast.Logging;
+using GLTFast.Materials;
 using Unity.Mathematics;
 using UnityEngine;
 
@@ -34,10 +35,12 @@ namespace GLTFast.Export {
         static readonly int k_EmissiveColor = Shader.PropertyToID("_EmissiveColor");
         static readonly int k_EmissionColorMap = Shader.PropertyToID("_EmissiveColorMap");
         static readonly int k_NormalMap = Shader.PropertyToID("_NormalMap");
+        static readonly int k_NormalScale = Shader.PropertyToID("_NormalScale");
         static readonly int k_BaseColorMap = Shader.PropertyToID("_BaseColorMap");
         static readonly int k_MaskMap = Shader.PropertyToID("_MaskMap");
         static readonly int k_SmoothnessRemapMax = Shader.PropertyToID("_SmoothnessRemapMax");
         static readonly int k_SmoothnessRemapMin = Shader.PropertyToID("_SmoothnessRemapMin");
+        static readonly int k_UnlitColor = Shader.PropertyToID("_UnlitColor");
 
         /// <summary>
         /// Converts a Unity material to a glTF material. 
@@ -58,7 +61,7 @@ namespace GLTFast.Export {
             };
 
             SetAlphaModeAndCutoff(uMaterial, material);
-            material.doubleSided = IsDoubleSided(uMaterial);
+            material.doubleSided = IsDoubleSided(uMaterial, MaterialGenerator.cullModePropId);
 
             //
             // Emission
@@ -104,7 +107,7 @@ namespace GLTFast.Export {
 
                 if (normalTex != null) {
                     if(normalTex is Texture2D) {
-                        material.normalTexture = ExportNormalTextureInfo(normalTex, uMaterial, gltf);
+                        material.normalTexture = ExportNormalTextureInfo(normalTex, uMaterial, gltf, k_NormalScale);
                         ExportTextureTransform(material.normalTexture, uMaterial, k_NormalMap, gltf);
                     } else {
                         logger?.Error(LogCode.TextureInvalidType, "normal", uMaterial.name );
@@ -141,32 +144,94 @@ namespace GLTFast.Export {
             var success = true;
             var pbr = new PbrMetallicRoughness { metallicFactor = 0, roughnessFactor = 1.0f };
 
+            var metallicUsed = false;
+            if (uMaterial.HasProperty(k_Metallic)) {
+                pbr.metallicFactor = uMaterial.GetFloat(k_Metallic);
+                metallicUsed = pbr.metallicFactor > 0;
+            }
+            
+            if (uMaterial.HasProperty(k_BaseColorMap)) {
+                // TODO if additive particle, render black into alpha
+                // TODO use private Material.GetFirstPropertyNameIdByAttribute here, supported from 2020.1+
+                var mainTex = uMaterial.GetTexture(k_BaseColorMap);
+
+                if (mainTex) {
+                    if(mainTex is Texture2D) {
+                        pbr.baseColorTexture = ExportTextureInfo(mainTex, gltf,
+                            material.alphaModeEnum == Material.AlphaMode.OPAQUE
+                                ? ImageExportBase.Format.Jpg
+                                : ImageExportBase.Format.Unknown
+                            );
+                        ExportTextureTransform(pbr.baseColorTexture, uMaterial, k_BaseColorMap, gltf);
+                    } else {
+                        logger?.Error(LogCode.TextureInvalidType, "main", uMaterial.name );
+                        success = false;
+                    }
+                }
+            }
+
             MaskMapImageExport ormImageExport = null;
             if (uMaterial.IsKeywordEnabled(k_KeywordMaskMap) && uMaterial.HasProperty(k_MaskMap)) {
                 var maskMap =  uMaterial.GetTexture(k_MaskMap) as Texture2D;
                 if (maskMap != null) {
-                    ormImageExport = new MaskMapImageExport(maskMap);
-                    if (AddImageExport(gltf, ormImageExport, out var ormTextureId)) {
-                        
-                        // TODO: smartly detect if metallic roughness channels are used and not create the
-                        // texture info if not. 
-                        pbr.metallicRoughnessTexture = new TextureInfo {
-                            index = ormTextureId
-                        };
-                        ExportTextureTransform(pbr.metallicRoughnessTexture, uMaterial, k_MaskMap, gltf);
-                        
-                        // TODO: smartly detect if occlusion channel is used and not create the
-                        // texture info if not.
-                        material.occlusionTexture = new OcclusionTextureInfo {
-                            index = ormTextureId
-                        };
-                        if (uMaterial.HasProperty(k_AORemapMin) ) {
-                            var occMin = uMaterial.GetFloat(k_AORemapMin);
-                            material.occlusionTexture.strength =  math.clamp(1-occMin,0,1);
+
+                    var smoothnessUsed = false;
+                    if (uMaterial.HasProperty(k_SmoothnessRemapMin)) {
+                        var smoothnessRemapMin = uMaterial.GetFloat(k_SmoothnessRemapMin);
+                        pbr.roughnessFactor = 1-smoothnessRemapMin;
+                        if (uMaterial.HasProperty(k_SmoothnessRemapMax)) {
+                            var smoothnessRemapMax = uMaterial.GetFloat(k_SmoothnessRemapMax);
+                            smoothnessUsed = math.abs(smoothnessRemapMin - smoothnessRemapMax) > math.EPSILON;
+                            if (smoothnessRemapMax < 1 && smoothnessUsed) {
+                                logger?.Warning(LogCode.RemapUnsupported,"Smoothness");
+                            }
+                        }
+                    }
+                    
+                    var occStrength = 1f;
+                    if (uMaterial.HasProperty(k_AORemapMin)) {
+                        var occMin = uMaterial.GetFloat(k_AORemapMin);
+                        occStrength =  math.clamp(1f-occMin,0,1);
+                        if (uMaterial.HasProperty(k_AORemapMax)) {
                             var occMax = uMaterial.GetFloat(k_AORemapMax);
-                            if (occMax < 1f) {
-                                // TODO: remap texture values
-                                logger?.Warning(LogCode.RemapUnsupported, "AO");
+                            if (occMax < 1f && occStrength > 0) {
+                                logger?.Warning(LogCode.RemapUnsupported,"AO");
+                            }
+                        }
+                    }
+
+                    var occUsed = occStrength > 0;
+                    
+                    // TODO: Detect if metallic/smoothness/occlusion channels
+                    // are used based on pixel values (i.e. have non-white
+                    // pixels) on top of parameter evaluation
+
+                    if ( metallicUsed || occUsed || smoothnessUsed ) {
+                        ormImageExport = new MaskMapImageExport(maskMap);
+                        if (AddImageExport(gltf, ormImageExport, out var ormTextureId)) {
+
+                            if (metallicUsed || smoothnessUsed) {
+                                pbr.metallicRoughnessTexture = new TextureInfo {
+                                    index = ormTextureId
+                                };
+                                ExportTextureTransform(pbr.metallicRoughnessTexture, uMaterial, k_MaskMap, gltf);
+                            }
+
+                            if (occStrength > 0) {
+                                // TODO: Detect if occlusion channel is used based
+                                // on pixel values
+                                // (i.e. have non-white pixels) and not assign the
+                                // texture info if not.
+                                material.occlusionTexture = new OcclusionTextureInfo {
+                                    index = ormTextureId,
+                                    strength = occStrength
+                                };
+                                ExportTextureTransform(
+                                    material.occlusionTexture,
+                                    uMaterial,
+                                    k_BaseColorMap, // HDRP Lit always re-uses baseColorMap transform
+                                    gltf
+                                );
                             }
                         }
                     }
@@ -181,38 +246,20 @@ namespace GLTFast.Export {
                 pbr.baseColor = uMaterial.GetColor(k_Color);
             }
 
-            if (uMaterial.HasProperty(k_BaseColorMap)) {
-                // TODO if additive particle, render black into alpha
-                // TODO use private Material.GetFirstPropertyNameIdByAttribute here, supported from 2020.1+
-                var mainTex = uMaterial.GetTexture(k_BaseColorMap);
-
-                if (mainTex) {
-                    if(mainTex is Texture2D) {
-                        pbr.baseColorTexture = ExportTextureInfo(mainTex, gltf);
-                        ExportTextureTransform(pbr.baseColorTexture, uMaterial, k_BaseColorMap, gltf);
-                    } else {
-                        logger?.Error(LogCode.TextureInvalidType, "main", uMaterial.name );
-                        success = false;
-                    }
-                }
-            }
-
-            if (uMaterial.HasProperty(k_Metallic)) {
-                pbr.metallicFactor = uMaterial.GetFloat(k_Metallic);
-            }
-
-            if (ormImageExport != null && uMaterial.HasProperty(k_SmoothnessRemapMax)) {
-                pbr.roughnessFactor = uMaterial.GetFloat(k_SmoothnessRemapMax);
-                if (uMaterial.HasProperty(k_SmoothnessRemapMin) && uMaterial.GetFloat(k_SmoothnessRemapMin) > 0) {
-                    logger?.Warning(LogCode.RemapUnsupported,"Smoothness");
-                }
-            } else
-            if(uMaterial.HasProperty(k_Smoothness)) {
+            if(ormImageExport == null && uMaterial.HasProperty(k_Smoothness)) {
                 pbr.roughnessFactor = 1f - uMaterial.GetFloat(k_Smoothness);
             }
 
             material.pbrMetallicRoughness = pbr;
             return success;
+        }
+        
+        protected override bool GetUnlitColor(UnityEngine.Material uMaterial, out Color baseColor) {
+            if (uMaterial.HasProperty(k_UnlitColor)) {
+                baseColor = uMaterial.GetColor(k_UnlitColor);
+                return true;
+            }
+            return base.GetUnlitColor(uMaterial, out baseColor);
         }
     }
 }