+++ Runtime/Scripts/Export/MeshMaterialCombination.cs
@@ -58,17 +58,24 @@ namespace GLTFast.Export {
 
         public override int GetHashCode() {
 #if NET_STANDARD
-            return HashCode.Combine(m_MeshId, m_MaterialIds);
-#else
-                var hash = 17;
-                hash = hash * 31 + m_MeshId.GetHashCode();
-                if (m_MaterialIds != null) {
-                    hash = hash * 31 + m_MaterialIds.Length;
-                    foreach (var id in m_MaterialIds) {
-                        hash = hash * 31 + id;
-                    }
+            var hash = new HashCode();
+            hash.Add(m_MeshId);
+            if (m_MaterialIds != null) {
+                foreach (var id in m_MaterialIds) {
+                    hash.Add(id);
                 }
-                return hash;
+            }
+            return hash.ToHashCode();
+#else
+            var hash = 17;
+            hash = hash * 31 + m_MeshId.GetHashCode();
+            if (m_MaterialIds != null) {
+                hash = hash * 31 + m_MaterialIds.Length;
+                foreach (var id in m_MaterialIds) {
+                    hash = hash * 31 + id;
+                }
+            }
+            return hash;
 #endif
         }
     }